make wallet root path
